Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/object-src/object-src-no-url-blocked-expected.txt

CONSOLE MESSAGE: Refused to load because it does not appear in the object-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load because it does not appear in the object-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

NOTRUN Should block the object and fire a spv
